requirements.txt
discord.py
ping3